requirements.txt
craiyon.py==0.1.0
openai==0.25.0
Pillow==9.3.0
requests==2.28.1
streamlit==1.16.0
streamlit_image_select==0.5.1
streamlit_lottie==0.0.3
~treamlit==1.15.2
